Split utils file and add num_layers_to_finetune to feature directories
