+++ tools/sense_studio/sense_studio.py
@@ -19,7 +19,8 @@ from flask import render_template
 from flask import request
 from flask import url_for
 
-from tools.sense_studio import utils
+from tools import utils
+from tools.sense_studio import utils as studio_utils
 from tools.sense_studio.annotation import annotation_bp
 from tools.sense_studio.video_recording import video_recording_bp
 
@@ -36,7 +37,7 @@ def projects_overview():
     Home page of SenseStudio. Show the overview of all registered projects and check if their
     locations are still valid.
     """
-    projects = utils.load_project_overview_config()
+    projects = studio_utils.load_project_overview_config()
 
     # Check if project paths still exist
     for name, project in projects.items():
@@ -50,7 +51,7 @@ def projects_list():
     """
     Provide the current list of projects to external callers.
     """
-    projects = utils.load_project_overview_config()
+    projects = studio_utils.load_project_overview_config()
     return jsonify(projects)
 
 
@@ -61,10 +62,10 @@ def project_config():
     """
     data = request.json
     name = data['name']
-    path = utils.lookup_project_path(name)
+    path = studio_utils.lookup_project_path(name)
 
     # Get config
-    config = utils.load_project_config(path)
+    config = studio_utils.load_project_config(path)
     return jsonify(config)
 
 
@@ -74,11 +75,11 @@ def remove_project(name):
     Remove a given project from the config file and reload the overview page.
     """
     name = urllib.parse.unquote(name)
-    projects = utils.load_project_overview_config()
+    projects = studio_utils.load_project_overview_config()
 
     del projects[name]
 
-    utils.write_project_overview_config(projects)
+    studio_utils.write_project_overview_config(projects)
 
     return redirect(url_for('projects_overview'))
 
@@ -114,7 +115,7 @@ def setup_project():
     # Update project config
     try:
         # Check for existing config file
-        config = utils.load_project_config(path)
+        config = studio_utils.load_project_config(path)
         old_name = config['name']
         config['name'] = name
     except FileNotFoundError:
@@ -126,7 +127,7 @@ def setup_project():
         }
         old_name = None
 
-    utils.write_project_config(path, config)
+    studio_utils.write_project_config(path, config)
 
     # Setup directory structure
     for split in utils.SPLITS:
@@ -135,7 +136,7 @@ def setup_project():
             os.mkdir(videos_dir)
 
     # Update overall projects config file
-    projects = utils.load_project_overview_config()
+    projects = studio_utils.load_project_overview_config()
 
     if old_name and old_name in projects:
         del projects[old_name]
@@ -144,7 +145,7 @@ def setup_project():
         'path': path,
     }
 
-    utils.write_project_overview_config(projects)
+    studio_utils.write_project_overview_config(projects)
 
     return redirect(url_for('project_details', project=name))
 
@@ -155,8 +156,8 @@ def project_details(project):
     Show the details for the selected project.
     """
     project = urllib.parse.unquote(project)
-    path = utils.lookup_project_path(project)
-    config = utils.load_project_config(path)
+    path = studio_utils.lookup_project_path(project)
+    config = studio_utils.load_project_config(path)
 
     stats = {}
     for class_name, tags in config['classes'].items():
@@ -178,15 +179,15 @@ def add_class(project):
     Add a new class to the given project.
     """
     project = urllib.parse.unquote(project)
-    path = utils.lookup_project_path(project)
+    path = studio_utils.lookup_project_path(project)
 
     # Get class name and tags
-    class_name, tag1, tag2 = utils.get_class_name_and_tags(request.form)
+    class_name, tag1, tag2 = studio_utils.get_class_name_and_tags(request.form)
 
     # Update project config
-    config = utils.load_project_config(path)
+    config = studio_utils.load_project_config(path)
     config['classes'][class_name] = [tag1, tag2]
-    utils.write_project_config(path, config)
+    studio_utils.write_project_config(path, config)
 
     # Setup directory structure
     for split in utils.SPLITS:
@@ -205,16 +206,16 @@ def edit_class(project, class_name):
     """
     project = urllib.parse.unquote(project)
     class_name = urllib.parse.unquote(class_name)
-    path = utils.lookup_project_path(project)
+    path = studio_utils.lookup_project_path(project)
 
     # Get new class name and tags
-    new_class_name, new_tag1, new_tag2 = utils.get_class_name_and_tags(request.form)
+    new_class_name, new_tag1, new_tag2 = studio_utils.get_class_name_and_tags(request.form)
 
     # Update project config
-    config = utils.load_project_config(path)
+    config = studio_utils.load_project_config(path)
     del config['classes'][class_name]
     config['classes'][new_class_name] = [new_tag1, new_tag2]
-    utils.write_project_config(path, config)
+    studio_utils.write_project_config(path, config)
 
     # Update directory names
     data_dirs = []
@@ -225,8 +226,12 @@ def edit_class(project, class_name):
             utils.get_tags_dir(path, split),
         ])
 
+        # Feature directories follow the format <dataset_dir>/<split>/<model>/<num_layers_to_finetune>/<label>
         features_dir = utils.get_features_dir(path, split)
-        data_dirs.extend([os.path.join(features_dir, model_dir) for model_dir in os.listdir(features_dir)])
+        model_dirs = [os.path.join(features_dir, model_dir) for model_dir in os.listdir(features_dir)]
+        data_dirs.extend([os.path.join(model_dir, tuned_layers)
+                          for model_dir in model_dirs
+                          for tuned_layers in os.listdir(model_dir)])
 
     logreg_dir = utils.get_logreg_dir(path)
     data_dirs.extend([os.path.join(logreg_dir, model_dir) for model_dir in os.listdir(logreg_dir)])
@@ -248,12 +253,12 @@ def remove_class(project, class_name):
     """
     project = urllib.parse.unquote(project)
     class_name = urllib.parse.unquote(class_name)
-    path = utils.lookup_project_path(project)
+    path = studio_utils.lookup_project_path(project)
 
     # Update project config
-    config = utils.load_project_config(path)
+    config = studio_utils.load_project_config(path)
     del config['classes'][class_name]
-    utils.write_project_config(path, config)
+    studio_utils.write_project_config(path, config)
 
     return redirect(url_for("project_details", project=project))
